requirements.txt
python-dotenv
selenium
pyautogui
selenium-stealth